Trying out River's collision code
I tried to add in the code for checking collisions that was on the CodeHS site. It doesn't seem to be working. I'll be trying out a structure similar to the one I used for my tower defense game, meaning I'll be making significant changes to current classes and adding in new classes(Sprite Superclass, collisionManager, etc.)